+++ land_consumption/core/operator_worker.py
@@ -2,11 +2,9 @@ import logging
 from typing import List
 
 import shapely
-from _duckdb import DuckDBPyConnection
-from climatoology.base.baseoperator import AoiProperties, BaseOperator, ComputationResources, Artifact
+from climatoology.base.baseoperator import AoiProperties, Artifact, BaseOperator, ComputationResources
 from climatoology.base.plugin_info import PluginInfo
 from ohsome import OhsomeClient
-from pyiceberg.catalog.rest import RestCatalog
 
 from land_consumption.components.land_consumption import get_land_consumption_artifacts
 from land_consumption.core.info import get_info
@@ -16,11 +14,9 @@ log = logging.getLogger(__name__)
 
 
 class LandConsumption(BaseOperator[ComputeInput]):
-    def __init__(self, data_connection: RestCatalog | tuple[RestCatalog, DuckDBPyConnection] | OhsomeClient):
+    def __init__(self, ohsome_client: OhsomeClient):
         super().__init__()
-        self.data_connection = data_connection
-        if isinstance(data_connection, RestCatalog) or isinstance(data_connection, tuple):
-            log.warning(f'Data connection {data_connection} is not tested')
+        self.ohsome_client = ohsome_client
 
         log.debug('Initialised Land consumption Operator')
 
@@ -37,6 +33,6 @@ class LandConsumption(BaseOperator[ComputeInput]):
         log.info(
             f'Handling compute request: {params.model_dump()} in region {aoi_properties.model_dump()} in context: {resources}'
         )
-        artifacts = get_land_consumption_artifacts(aoi, data_connection=self.data_connection, resources=resources)
+        artifacts = get_land_consumption_artifacts(aoi, ohsome_client=self.ohsome_client, resources=resources)
 
         return artifacts